used in the unpacking queue

+++ src/pynxtools_em/examples/get_file_from_archive_formats.py
@@ -40,14 +40,15 @@ def get_file_from_zip(
         target_dir = Path(target_directory)
         target_dir.mkdir(parents=True, exist_ok=True)
         target_file = target_dir / target_file_name
-        with zipfile.ZipFile(zip_file_path, "r") as zf:
-            # target_file.write_bytes(zf.read(file_in_zip_path))
-            with (
-                zf.open(file_in_zip_path) as src,
-                open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
-            ):
-                shutil.copyfileobj(src, dst)
-        return os.path.isfile(target_file)
+        if not os.path.isfile(target_file):
+            with zipfile.ZipFile(zip_file_path, "r") as zf:
+                # target_file.write_bytes(zf.read(file_in_zip_path))
+                with (
+                    zf.open(file_in_zip_path) as src,
+                    open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
+                ):
+                    shutil.copyfileobj(src, dst)
+            return os.path.isfile(target_file)
     except (FileNotFoundError, KeyError, zipfile.BadZipFile) as exception:
         print(f"Error extracting file from zip: {exception}")
     return False
@@ -63,16 +64,17 @@ def get_file_from_tar(
         target_dir = Path(target_directory)
         target_dir.mkdir(parents=True, exist_ok=True)
         target_file = target_dir / target_file_name
-        with tarfile.open(tar_file_path, mode="r:*") as tf:
-            member = tf.getmember(file_in_tar_path)
-            if not member.isfile():
-                raise ValueError(f"{file_in_tar_path} is not a regular file")
-            with (
-                tf.extractfile(member) as src,
-                open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
-            ):
-                shutil.copyfileobj(src, dst)
-        return os.path.isfile(target_file)
+        if not os.path.isfile(target_file):
+            with tarfile.open(tar_file_path, mode="r:*") as tf:
+                member = tf.getmember(file_in_tar_path)
+                if not member.isfile():
+                    raise ValueError(f"{file_in_tar_path} is not a regular file")
+                with (
+                    tf.extractfile(member) as src,
+                    open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
+                ):
+                    shutil.copyfileobj(src, dst)
+            return os.path.isfile(target_file)
     except (
         FileNotFoundError,  # tar file missing or target path invalid
         KeyError,  # member not found in archive
@@ -93,16 +95,17 @@ def get_file_from_rar(
         target_dir = Path(target_directory)
         target_dir.mkdir(parents=True, exist_ok=True)
         target_file = target_dir / target_file_name
-        with rarfile.RarFile(rar_file_path) as rf:
-            info = rf.getinfo(file_in_rar_path)
-            if info.isdir():
-                raise ValueError(f"{file_in_rar_path} is a directory")
-            with (
-                rf.open(info) as src,
-                open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
-            ):
-                shutil.copyfileobj(src, dst)
-        return os.path.isfile(target_file)
+        if not os.path.isfile(target_file):
+            with rarfile.RarFile(rar_file_path) as rf:
+                info = rf.getinfo(file_in_rar_path)
+                if info.isdir():
+                    raise ValueError(f"{file_in_rar_path} is a directory")
+                with (
+                    rf.open(info) as src,
+                    open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
+                ):
+                    shutil.copyfileobj(src, dst)
+            return os.path.isfile(target_file)
     except (
         FileNotFoundError,
         KeyError,
@@ -123,14 +126,20 @@ def get_file_from_sevenzip(
         target_dir = Path(target_directory)
         target_dir.mkdir(parents=True, exist_ok=True)
         target_file = target_dir / target_file_name
-        with py7zr.SevenZipFile(sevenz_file_path, mode="r") as z:
-            # get a file-like object for streaming
-            with (
-                z.open(file_in_sevenz_path) as src,
-                open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
-            ):
-                shutil.copyfileobj(src, dst)
-        return os.path.isfile(target_file)
+        if not os.path.isfile(target_file):
+            with py7zr.SevenZipFile(sevenz_file_path, mode="r") as z:
+                # get a file-like object for streaming
+                # with (
+                #     z.open(file_in_sevenz_path) as src,
+                #     open(target_file, "wb", buffering=BUFFER_SIZE) as dst,
+                # ):
+                #     shutil.copyfileobj(src, dst)
+                # in memory extraction
+                extracted = z.read([file_in_sevenz_path])
+                src = extracted[file_in_sevenz_path]  # BytesIO object
+                with open(target_file, "wb", buffering=BUFFER_SIZE) as dst:  # streaming
+                    shutil.copyfileobj(src, dst, length=BUFFER_SIZE)
+            return os.path.isfile(target_file)
     except (
         FileNotFoundError,
         KeyError,
